refactor(capabilities): make capabilities a module folder; home `your` there
Convert `src/capabilities.rs` into a `src/capabilities/` module folder so
each capability can get its own file, and move the personalization layer to
`src/capabilities/your.rs`. The smaller host capabilities stay inline in
`mod.rs` for now and can be split out incrementally.

Also retarget the `your` roadmap in specs/your.md onto everruns' existing
declarative-capability mechanism (`DeclarativeCapabilityDefinition`, id
`declarative:<name>`) — which already contributes a system prompt, mounted
files, skills, and MCP servers from data — instead of a bespoke manifest.
Global skills then fall out as a declarative capability's skills mounted at
the standard skills discovery path.
